Fix module.yaml and move benchmarks to corresponding folder.
